requirements.txt
Flask>=2.0
google-generativeai>=0.9
PyMuPDF>=1.22.0
Pillow>=9.0.0
